Social media app backend

+++ SocialMediaApp/src/main/java/com/example/lasya/SocialMediaApp/controller/CommentController.java
@@ -1,24 +1,43 @@
 package com.example.lasya.SocialMediaApp.controller;
 
+import com.example.lasya.SocialMediaApp.bean.PostBean;
 import com.example.lasya.SocialMediaApp.entity.Comment;
+import com.example.lasya.SocialMediaApp.exception.CommentNotFoundException;
+import com.example.lasya.SocialMediaApp.repository.CommentRepository;
 import com.example.lasya.SocialMediaApp.service.CommentService;
+import com.example.lasya.SocialMediaApp.service.PostService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Map;
 
+@CrossOrigin
+@RestController
 public class CommentController {
     private final CommentService commentService;
 
+    private final PostService postService;
+
+    private final CommentRepository commentRepository;
+
+    private static final Logger logger = LoggerFactory.getLogger(CommentController.class);
+
     @Autowired
-    public CommentController(CommentService commentService) {
+    public CommentController(CommentService commentService, PostService postService, CommentRepository commentRepository) {
         super();
         this.commentService = commentService;
+        this.postService = postService;
+        this.commentRepository = commentRepository;
     }
 
     @ApiOperation(value = "This API is used to get a comment based on a given postId")
@@ -37,18 +56,22 @@ public class CommentController {
     }
 
     @ApiOperation(value = "This API is used to delete a comment based on a given commentId")
-    @DeleteMapping("/api/v1/comment/{commentId}")
+    @DeleteMapping("/api/v1/comment/delete/{commentId}")
     public ResponseEntity<String> deleteCommentById(@PathVariable int commentId) {
+        if(!commentRepository.existsByCommentId(commentId)){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Comment not found with the commentId: " + commentId);
+        }
         commentService.deleteByCommentId(commentId);
         return ResponseEntity.ok("Comment deleted successfully");
     }
 
     @ApiOperation(value = "This API is used to edit a comment based on a given commentId")
-    @PutMapping("/api/v1/comment/{commentId}")
+    @PutMapping("/api/v1/comment/edit/{commentId}")
     public ResponseEntity<String> editComment(
             @PathVariable int commentId,
             @RequestBody Map<String, String> request) {
-        String newText = request.get("newText");
+        String newText = request.get("text");
+        logger.info("commentId: {}, newText: {}", commentId, newText);
         commentService.editComment(commentId, newText);
         return ResponseEntity.ok("Comment edited successfully");
     }
@@ -59,7 +82,23 @@ public class CommentController {
             @PathVariable int postId,
             @RequestBody Map<String, String> request) {
         String text = request.get("text");
-        commentService.addCommentToPost(postId, text);
+        int userId = Integer.parseInt(request.get("userId"));
+        String parentCommentIdStr = request.get("parentCommentId");
+        Integer parentCommentId = null;
+        if (parentCommentIdStr != null) {
+            parentCommentId = Integer.parseInt(parentCommentIdStr);
+        }
+        String uploadTimeString = request.get("uploadTime");
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+        LocalDateTime uploadTime = LocalDateTime.parse(uploadTimeString, formatter);
+
+        // Check if the postId corresponds to an existing post in the database
+        PostBean checkIfIdExists = postService.getPostById(postId);
+        logger.info(String.valueOf(checkIfIdExists));
+        if (postService.getPostById(postId) == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Post not found for postId: " + postId);
+        }
+        commentService.addCommentToPost(postId, userId, text, parentCommentId, uploadTime);
         return ResponseEntity.ok("Comment added successfully");
     }
